enable clocks from startup script
Removed old startup code from model, use samd-peripherals startup code
