+++ source/codal_target_hal.cpp
@@ -24,6 +24,7 @@ DEALINGS IN THE SOFTWARE.
 #include "codal_target_hal.h"
 #include "Timer.h"
 #include "CodalDmesg.h"
+#include "CodalCompat.h"
 
 void target_enable_irq()
 {
@@ -38,6 +39,16 @@ void target_disable_irq()
 void target_wait_for_event()
 {
     __WFE();
+}
+
+int target_seed_random(uint32_t rand)
+{
+    return codal::seed_random(rand);
+}
+
+int target_random(int max)
+{
+    return codal::random(max);
 }
 
 void target_wait_us(unsigned long us) {
